accessible/tests/mochitest/textcaret/test_general.html

<html>

<head>
  <title>Text accessible caret testing</title>

  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">
    function turnCaretBrowsing(aIsOn) {
      Services.prefs.setBoolPref("accessibility.browsewithcaret", aIsOn);
    }

    function testCaretOffset(aID, aCaretOffset) {
      const acc = getAccessible(aID, [nsIAccessibleText]);
      is(acc.caretOffset, aCaretOffset,
         "Wrong caret offset for " + aID);
    }

    function testCaretOffsets(aList) {
      for (let i = 0; i < aList.length; i++)
        testCaretOffset(aList[0][0], aList[0][1]);
    }

    async function doSetCaretOffset(aID, aOffset, aFocusTargetID) {
      let target = getAccessible(aID, [nsIAccessibleText]);
      let offset = aOffset == -1 ? target.characterCount : aOffset;
      let focus = aFocusTargetID ? getAccessible(aFocusTargetID) : null;

      let events = [[EVENT_TEXT_CARET_MOVED, target]];
      if (focus) {
        events.push([EVENT_FOCUS, focus]);
      }
      let p = waitForEvents(events);
      target.caretOffset = offset;
      let [caretEvt] = await p;
      caretEvt.QueryInterface(nsIAccessibleCaretMoveEvent);
      is(caretEvt.caretOffset, offset,
         "Wrong caret offset for " + prettyName(aID));
      is(caretEvt.isSelectionCollapsed, true,
         "Wrong collapsed for " + prettyName(aID));
    }

    async function doMoveCaretToDOMPoint(aID, aDOMPointNode, aDOMPointOffset,
                                         aExpectedOffset, aFocusTargetID,
                                         aCheckFunc) {
      let target = getAccessible(aID, [nsIAccessibleText]);
      let focus = aFocusTargetID ? getAccessible(aFocusTargetID) : null;
      let focusNode = focus ? focus.DOMNode : null;

      let events = [[EVENT_TEXT_CARET_MOVED, target]];
      if (focus) {
        events.push([EVENT_FOCUS, focus]);
      }
      let p = waitForEvents(events);

      if (focusNode) {
        focusNode.focus();
      }

      let selection = aDOMPointNode.documentGlobal.getSelection();
      let selRange = selection.rangeCount
        ? selection.getRangeAt(0)
        : aDOMPointNode.ownerDocument.createRange();
      selRange.setStart(aDOMPointNode, aDOMPointOffset);
      selRange.collapse(true);
      selection.removeAllRanges();
      selection.addRange(selRange);

      let [caretEvt] = await p;
      caretEvt.QueryInterface(nsIAccessibleCaretMoveEvent);
      is(caretEvt.caretOffset, aExpectedOffset,
         "Wrong caret offset for " + prettyName(aID));
      is(caretEvt.isSelectionCollapsed, true,
         "Wrong collapsed for " + prettyName(aID));

      if (aCheckFunc) {
        aCheckFunc.call();
      }
    }

    async function doTests() {
      turnCaretBrowsing(true);

      // test caret offsets
      testCaretOffset(document, 0);  // because of no selection ranges
      testCaretOffset("textbox", -1);
      testCaretOffset("textarea", -1);
      testCaretOffset("p", -1);

      // test caret move events and caret offsets
      let focusEvt = waitForEvent(EVENT_FOCUS, "textbox");
      getNode("textbox").focus();
      await focusEvt;
      testCaretOffset("textbox", 0);

      await doSetCaretOffset("textbox", 1);
      await doSetCaretOffset("link", 1, "link");
      await doSetCaretOffset("heading", 1, document);

      // a*b*c
      let p2Doc = getNode("p2_container").contentDocument;
      let traversalList = [
        { // before 'a'
          DOMPoint: [ getNode("p2", p2Doc).firstChild, 0 ],
          point: [ getNode("p2", p2Doc), 0 ],
          pointList: [ [ p2Doc, 0 ] ],
        },
        { // after 'a' (before anchor)
          DOMPoint: [ getNode("p2", p2Doc).firstChild, 1 ],
          point: [ getNode("p2", p2Doc), 1 ],
          pointList: [ [ p2Doc, 0 ] ],
        },
        { // before 'b' (inside anchor)
          DOMPoint: [ getNode("p2_a", p2Doc).firstChild, 0 ],
          point: [ getNode("p2_a", p2Doc), 0 ],
          pointList: [
            [ getNode("p2", p2Doc), 1 ],
            [ p2Doc, 0 ],
          ],
        },
        { // after 'b' (inside anchor)
          DOMPoint: [ getNode("p2_a", p2Doc).firstChild, 1 ],
          point: [ getNode("p2_a", p2Doc), 1 ],
          pointList: [
            [ getNode("p2", p2Doc), 1 ],
            [ p2Doc, 0 ],
          ],
        },
        { // before 'c' (after anchor)
          DOMPoint: [ getNode("p2", p2Doc).lastChild, 0 ],
          point: [ getNode("p2", p2Doc), 2 ],
          pointList: [ [ p2Doc, 0 ] ],
        },
        { // after 'c'
          DOMPoint: [ getNode("p2", p2Doc).lastChild, 1 ],
          point: [ getNode("p2", p2Doc), 3 ],
          pointList: [ [ p2Doc, 0 ] ],
        },
      ];

      const focusNode = getNode("p2", p2Doc);
      for (let i = 0; i < traversalList.length; i++) {
        const item = traversalList[i];
        const domPointNode = item.DOMPoint[0];
        const domPointOffset = item.DOMPoint[1];
        const textAcc = item.point[0];
        const textOffset = item.point[1];
        const textList = item.pointList;
        const aFocusTargetID = i == 0 ? focusNode : null;
        await doMoveCaretToDOMPoint(
          textAcc, domPointNode, domPointOffset, textOffset,
          aFocusTargetID,
          testCaretOffsets.bind(null, textList)
        );
      }

      turnCaretBrowsing(false);
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTests);
  </script>
</head>

<body>

  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=448744"
     title="caretOffset should return -1 if the system caret is not currently with in that particular object">
    Bug 448744
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=524115"
     title="HyperText accessible should get focus when the caret is positioned inside of it, text is changed or copied into clipboard by ATs">
    Bug 524115
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=546068"
     title="Position is not being updated when atk_text_set_caret_offset is used">
    Bug 546068
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=672717"
     title="Broken caret when moving into/out of embedded objects with right arrow">
    Bug 672717
  </a>
  <a target="_blank"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=725581"
     title="caretOffset for textarea should be -1 when textarea doesn't have a focus">
    Bug 725581
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <input id="textbox" value="hello"/>
  <textarea id="textarea">text<br>text</textarea>
  <p id="p" contentEditable="true"><span>text</span><br/>text</p>
  <a id="link" href="about:mozilla">about mozilla</a>
  <h5 id="heading">heading</h5>
  <iframe id="p2_container"
          src="data:text/html,<p id='p2' contentEditable='true'>a<a id='p2_a' href='mozilla.org'>b</a>c</p>"></iframe>

  <div id="eventdump"></div>
</body>
</html>
